PLATUI-2717: Converted to multi-project build, disabled VNU due to issues with validator library
