+++ project/LibDependencies.scala
@@ -17,22 +17,36 @@
 import sbt._
 
 object LibDependencies {
-  val all: Seq[ModuleID] = Seq(
-    "com.typesafe.play"            %% "play-json"             % "2.8.1",
-    "org.pegdown"                   % "pegdown"               % "1.6.0",
-    "org.scalatest"                %% "scalatest"             % "3.2.9",
-    "com.typesafe"                  % "config"                % "1.4.1",
-    "com.beachape"                 %% "enumeratum"            % "1.7.0",
-    "nu.validator"                  % "validator"             % "20.7.2",
-    "com.softwaremill.magnolia1_2" %% "magnolia"              % "1.1.2",
-    "org.reflections"               % "reflections"           % "0.10.2",
-    "org.scalacheck"               %% "scalacheck"            % "1.14.0",
-    "org.scalatestplus.play"       %% "scalatestplus-play"    % "5.1.0",
-    "com.typesafe.play"            %% "play-test"             % "2.8.16",
-    "com.typesafe.play"            %% "filters-helpers"       % "2.8.16",
-    "org.scalatestplus"            %% "scalatestplus-mockito" % "1.0.0-M2"       % "test",
-    "com.vladsch.flexmark"          % "flexmark-all"          % "0.35.10"        % "it,test",
-    "com.softwaremill.diffx"       %% "diffx-scalatest"       % "0.5.6"          % "it,test",
-    "uk.gov.hmrc"                  %% "play-frontend-hmrc"    % "7.19.0-play-28" % "it,test"
+
+  val play28 = play("play-28")
+  val play29 = play("play-29")
+  val play30 = play("play-30")
+
+  private def play(playSuffix: String) = Seq(
+    "org.scalatest"                %% "scalatest"                       % "3.2.17",
+    "com.beachape"                 %% "enumeratum"                      % "1.7.0",
+    "com.softwaremill.magnolia1_2" %% "magnolia"                        % "1.1.2",
+    "org.reflections"               % "reflections"                     % "0.10.2",
+    "org.scalacheck"               %% "scalacheck"                      % "1.14.0",
+    "org.scalatestplus.play"       %% "scalatestplus-play"              % scalaTestPlusPlayVersion(playSuffix),
+    "nu.validator"                  % "validator"                       % "20.7.2",
+    "org.scalatestplus"            %% "scalatestplus-mockito"           % "1.0.0-M2"                     % "test",
+    "com.softwaremill.diffx"       %% "diffx-scalatest"                 % "0.5.6"                        % "test",
+    "com.vladsch.flexmark"          % "flexmark-all"                    % flexmarkAllVersion(playSuffix) % "test",
+    "uk.gov.hmrc"                  %% s"play-frontend-hmrc-$playSuffix" % "8.3.0"                        % "test"
   )
+
+  private def scalaTestPlusPlayVersion(playSuffix: String): String =
+    playSuffix match {
+      case "play-28" => "5.1.0"
+      case "play-29" => "6.0.0"
+      case "play-30" => "7.0.0"
+    }
+
+  private def flexmarkAllVersion(playSuffix: String): String =
+    playSuffix match {
+      case "play-28" => "0.62.2"
+      case "play-29" => "0.64.8"
+      case "play-30" => "0.64.8"
+    }
 }
