+++ scalatest-accessibility-linter-play-30/src/main/resources/uk/gov/hmrc/scalatestaccessibilitylinter/accessibility-linter.conf
