+++ scalatest-accessibility-linter-play-30/src/main/scala/uk/gov/hmrc/scalatestaccessibilitylinter/config.scala
