+++ scalatest-accessibility-linter-play-30/src/test/scala/uk/gov/hmrc/scalatestaccessibilitylinter/views/AutomaticAccessibilitySpecSpec.scala
